Give at least a way how to increase entropy on VMs.
On virtualboxes entropy is usually harder to get. This makes generating
GPG keys a very long task. We can shorten this process, but I am not
sure about the consequences security-wise. It is normally a good idea to
use /dev/urandom (instead of /dev/random), but this might be one of the
rare exceptions.